Added: GODMODE
fixed various warnings

+++ Goodbros/Enemy.cpp
@@ -3,7 +3,7 @@
 #include "ModuleCollision.h"
 #include "ModuleRender.h"
 
-Enemy::Enemy(int x, int y) : position(x, y), collider(nullptr)
+Enemy::Enemy(float x, float y) : position(x, y), collider(nullptr)
 {}
 
 Enemy::~Enemy()
