+++ Goodbros/Enemy_Cowboy_Blue.cpp
@@ -3,7 +3,7 @@
 #include "ModuleCollision.h"
 #include "ModulePlayer.h"
 
-Enemy_CowBoy_Blue::Enemy_CowBoy_Blue(int x, int y) : Enemy(x, y)
+Enemy_CowBoy_Blue::Enemy_CowBoy_Blue(float x, float y) : Enemy(x, y)
 {
 	forward.PushBack({ 117, 396, 30, 55 });
 	forward.PushBack({ 152, 396, 30, 55 });
