+++ Goodbros/Enemy_Cowboy_Green.h
@@ -20,7 +20,7 @@ private:
 	iPoint original_pos;
 public:
 
-	Enemy_CowBoy_Green(int x, int y);
+	Enemy_CowBoy_Green(float x, float y);
 	void Die();
 	void Move();
 };
