+++ Goodbros/ModuleParticles.h
@@ -41,7 +41,7 @@ public:
 	bool CleanUp();
 	void OnCollision(Collider* c1, Collider* c2);
 
-	void AddParticle(const Particle& particle, int x, int y, COLLIDER_TYPE collider_type, Uint32 delay);
+	void AddParticle(const Particle& particle, float x, float y, COLLIDER_TYPE collider_type, Uint32 delay);
 private:
 
 	SDL_Texture* graphics = nullptr;
